Add a resolve_binding method to Context
