+++ crates/ruff/resources/test/fixtures/pylint/sys_exit_alias_3.py
@@ -7,3 +7,10 @@ quit(0)
 def main():
     exit(1)
     quit(1)
+
+
+def main():
+    exit = 1
+
+    exit(1)
+    quit(1)
