+++ crates/ruff/src/rules/pylint/rules/sys_exit_alias.rs
@@ -2,8 +2,6 @@ use rustpython_parser::ast::{Expr, ExprKind};
 
 use ruff_diagnostics::{AutofixKind, Diagnostic, Edit, Violation};
 use ruff_macros::{derive_message_formats, violation};
-use ruff_python_ast::context::Context;
-use ruff_python_ast::scope::{BindingKind, FromImportation, Importation, SubmoduleImportation};
 use ruff_python_ast::types::Range;
 
 use crate::checkers::ast::Checker;
@@ -28,52 +26,6 @@ impl Violation for SysExitAlias {
     }
 }
 
-/// Return the appropriate `sys.exit` reference based on the current set of
-/// imports, or `None` is `sys.exit` hasn't been imported.
-fn get_member_import_name_alias(context: &Context, module: &str, member: &str) -> Option<String> {
-    context.scopes().find_map(|scope| {
-        scope
-            .binding_ids()
-            .find_map(|index| match &context.bindings[*index].kind {
-                // e.g. module=sys object=exit
-                // `import sys`         -> `sys.exit`
-                // `import sys as sys2` -> `sys2.exit`
-                BindingKind::Importation(Importation { name, full_name }) => {
-                    if full_name == &module {
-                        Some(format!("{name}.{member}"))
-                    } else {
-                        None
-                    }
-                }
-                // e.g. module=os.path object=join
-                // `from os.path import join`          -> `join`
-                // `from os.path import join as join2` -> `join2`
-                BindingKind::FromImportation(FromImportation { name, full_name }) => {
-                    let mut parts = full_name.split('.');
-                    if parts.next() == Some(module)
-                        && parts.next() == Some(member)
-                        && parts.next().is_none()
-                    {
-                        Some((*name).to_string())
-                    } else {
-                        None
-                    }
-                }
-                // e.g. module=os.path object=join
-                // `import os.path ` -> `os.path.join`
-                BindingKind::SubmoduleImportation(SubmoduleImportation { full_name, .. }) => {
-                    if full_name == &module {
-                        Some(format!("{full_name}.{member}"))
-                    } else {
-                        None
-                    }
-                }
-                // Non-imports.
-                _ => None,
-            })
-    })
-}
-
 /// PLR1722
 pub fn sys_exit_alias(checker: &mut Checker, func: &Expr) {
     let ExprKind::Name { id, .. } = &func.node else {
@@ -93,9 +45,9 @@ pub fn sys_exit_alias(checker: &mut Checker, func: &Expr) {
             Range::from(func),
         );
         if checker.patch(diagnostic.kind.rule()) {
-            if let Some(content) = get_member_import_name_alias(&checker.ctx, "sys", "exit") {
+            if let Some(binding) = checker.ctx.resolve_qualified_import_name("sys", "exit") {
                 diagnostic.set_fix(Edit::replacement(
-                    content,
+                    binding,
                     func.location,
                     func.end_location.unwrap(),
                 ));
